'Add to favorites' feature implemented.
If stop does not exist in the db file (favorite stops) on the device, then it gets added. Otherwise, the user gets notified that
the stop already exists.

Added context menu for each bus stop recycler view item for the 'add to favorites' functionality.

The "three-dot" overflow pattern was used as icon. Adding a bus stop to favorites needs to be programmed.