New feature added - pull to refresh.
It uses SwipeRefreshLayout and is used to update bus arrivals at a given stop.

Content sharing feature implemented. Departure information of selected route can be shared with others.
Fixed issue that prevented from viewing favorites and transit alerts after searching for a bus stop.

Transit alerts feature implemented.

User can see a list of transit alerts for King County Metro that comes from an RSS feed. The list contains a title, as well as a
link that can be followed in order to get detailed information on the selected alert.

'View favorite stops' feature implemented.

Favorite stops can be viewed by tapping on the 'Favorite stops' menu item. The user can also remove a given stop from their
list of favorites by tapping on the three dots (next to the stop id). The RecyclerViewAdapter gets notified of the dataset changes
that were made, so once the stop has been deleted, the change gets reflected instantly.

'Add to favorites' feature implemented.

If stop does not exist in the db file (favorite stops) on the device, then it gets added. Otherwise, the user gets notified that
the stop already exists.

Added context menu for each bus stop recycler view item for the 'add to favorites' functionality.

The "three-dot" overflow pattern was used as icon. Adding a bus stop to favorites needs to be programmed.